fix: 프로젝트 버그 수정 및 Member 버그 수정

+++ src/main/java/Project/Teaming/Invite/Dto/AcceptInviteRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class AcceptInviteRequestDto {
 
-    private Long inviteId;
+    private Integer inviteId;
 }
 
 // test committ
